move `showToast` to `utils`
